main8

+++ playground2/main.go
@@ -11,5 +11,5 @@ func main() {
 	// fmt.Println(pi)
 
 	const myConst int = 1
-	fmt.Printf("type: %T, value: %v\n", myConst+1.1, myConst+1.1)
+	fmt.Printf("type: %T, value: %v\n", float64(myConst)+1.1, float64(myConst)+1.1)
 }
